moved the client part outside
