+++ insights-ui/src/components/projects/ProjectSummaryCard.tsx
@@ -1,6 +1,3 @@
-'use client';
-import { ProjectDetails } from '@/types/project/project';
-import { useRouter } from 'next/navigation';
 import Thumbnail from '@dodao/web-core/components/app/Thumbnail';
 import Card from '@dodao/web-core/components/core/card/Card';
 import { shorten } from '@dodao/web-core/utils/utils';
@@ -10,6 +7,7 @@ import { InsightsConstants } from '@/util/insights-constants';
 import { regenerateReport } from '@/util/regenerate';
 import EllipsisDropdown from '@dodao/web-core/components/core/dropdowns/EllipsisDropdown';
 import { isAdmin } from '@/util/auth/isAdmin';
+import Elipsis from './Elipsis';
 
 interface ProjectSummaryCardProps {
   projectId: string;
@@ -20,30 +18,6 @@ const MODEL_OPTIONS = [
 ];
 
 const ProjectSummaryCard: React.FC<ProjectSummaryCardProps> = ({ projectId }) => {
-  const router = useRouter();
-
-  const tableActions = {
-    items: [
-      { key: 'view', label: 'View' },
-      { key: 'edit', label: 'Edit' },
-      ...MODEL_OPTIONS.map((model) => ({
-        key: `regenerate_${model.key}`,
-        label: `Regenerate with ${model.label}`,
-      })),
-    ],
-    onSelect: async (key: string, item: string) => {
-      if (key === 'view') {
-        router.push(`/crowd-funding/projects/${item}`);
-      } else if (key === 'edit') {
-        router.push(`/crowd-funding/projects/${item}/edit`);
-      } else if (key.startsWith('regenerate_')) {
-        const model = key.replace('regenerate_', '');
-        const { success, message } = await regenerateReport(item, model);
-        success ? router.refresh() : alert(message);
-      }
-    },
-  };
-
   function formatProjectName(projectId: string): string {
     return projectId
       .trim() // Remove extra spaces
@@ -55,16 +29,7 @@ const ProjectSummaryCard: React.FC<ProjectSummaryCardProps> = ({ projectId }) =>
     <Card>
       <Link href={`/crowd-funding/projects/${projectId}`} className="card blog-card w-inline-block h-full w-full">
         <div className="relative w-full">
-          {
-            <div className="absolute top-2 right-2 py-2 pl-3 pr-4 text-right font-medium sm:pr-0">
-              <EllipsisDropdown
-                items={tableActions.items}
-                onSelect={(key) => {
-                  tableActions.onSelect(key, projectId);
-                }}
-              />
-            </div>
-          }
+          {isAdmin() && <Elipsis projectId={projectId} />}
 
           <Thumbnail
             src={`https://${InsightsConstants.S3_BUCKET_NAME}.s3.us-east-1.amazonaws.com/images/${projectId}`}
